feat(ci): carve out storage layout CI
Since the storage layout CI needs the API keys to run, it does not work
in PRs originating from forked repos. To avoid those unnecessary errors,
this PR splits the workflow into two parts: the first that generates the
compiled layouts and the second that fetches the deployed layouts and
compares them.

Such a split is possible because the second type of workflow runs in a
trusted context, and hence, the API key is accessible to it. However,
testing this setup is always difficult so I will test it out on my
repository before sending it out.

The testing is currently blocked by a PR that I sent to foundry, to
allow dumping the deployed state layout JSON to disk.