fix(ci): use PR addrs to compare layouts
The `compare-layouts.yml` workflow runs via `workflow_run` and is thus
triggered in the context of the base branch. It is, therefore, reading
potentially outdated information from the `deployedContract.json` file.
To avoid that, we use an intermediate artifact generated from the PR,
which contains the updated addresses, during the step that fetches the
storage layouts from Etherscan.

Simpler validation, such as all required addresses are present and
correctly formatted, is retained in the parent CI to permit quicker
short circuit upon receiving invalid data.